refactor: remove tight coupling between plugins (utc, advancedParse)
make core module independent of utc plugin;
prepare core module and plugin utc to avoid code duplication in
plugin advancedParse.
